[OG-65] Bug fixes
