+++ polymorphia-frontend/components/equipment/modals/OpeningChestModalContent.tsx
@@ -3,7 +3,7 @@ import { EquipmentContext } from "@/components/providers/EquipmentContext";
 import { API_STATIC_URL } from "@/services/api";
 import { useContext } from "react";
 import Image from "next/image";
-import { useModal } from "@/components/modal/ModalContext";
+import { useModal } from "@/components/providers/ModalContext";
 
 export default function OpeningChestModalContent() {
   const { closeModal } = useModal();
@@ -14,7 +14,6 @@ export default function OpeningChestModalContent() {
 
   const handlePickItem = (itemId: number) => {
     if (!openingChest) return;
-
     const isPicked = pickedItemsIds.includes(itemId);
 
     if (openingChest.behavior === "ONE_OF_MANY") {
@@ -29,28 +28,26 @@ export default function OpeningChestModalContent() {
   return (
     <>
       <div className="w-80 grid grid-cols-2 gap-3">
-        {openingChest?.items.map((item) => {
-          const isPicked = pickedItemsIds.includes(item.itemId);
-
-          return (
-            <div
-              key={item.itemId}
-              className="relative w-full aspect-square"
-              onClick={() => handlePickItem(item.itemId)}
-            >
-              <Image
-                src={`${API_STATIC_URL}/${item.imageUrl}`}
-                alt={item.title}
-                fill
-                className={`equipment-img ${
-                  isPicked ? "outline-4 outline-amber-400" : ""
-                }`}
-                priority
-                sizes="(min-width: 1024px) 10vw, 25vw"
-              />
-            </div>
-          );
-        })}
+        {openingChest?.items.map(item => (
+          <div
+            key={item.itemId}
+            className="relative w-full aspect-square cursor-pointer"
+            onClick={() => handlePickItem(item.itemId)}
+          >
+            <Image
+              src={`${API_STATIC_URL}/${item.imageUrl}`}
+              alt={item.title}
+              fill
+              className={`equipment-img ${
+                pickedItemsIds.includes(item.itemId)
+                  ? "outline-4 outline-amber-400"
+                  : ""
+              }`}
+              priority
+              sizes="(min-width: 1024px) 10vw, 25vw"
+            />
+          </div>
+        ))}
       </div>
       <div className="w-full mt-5">
         <ButtonWithBorder
