+++ polymorphia-frontend/components/modal/Modal.tsx
@@ -1,12 +1,20 @@
-import { ModalProps } from "@/interfaces/modal/ModalInterfaces";
-import ModalInner from "./ModalInner";
-import { useState, useEffect } from "react";
-import { ModalProvider } from "./ModalContext";
+"use client"
+
+import clsx from "clsx";
+import { X } from "lucide-react";
+import { createPortal } from "react-dom";
+import "../../styles/modal.css";
+import { useAnimatedModalState } from "@/animations/Modal";
+import {useEffect, useRef, useState} from "react";
+import { ModalProvider } from "../providers/ModalContext";
+import {ModalProps} from "@/interfaces/modal/ModalInterfaces";
 
 export default function Modal(props: ModalProps) {
-  const { isDataPresented, onClosed, ...rest } = props;
+  const { isDataPresented, onClosed, title, subtitle, children, ...rest } = props;
 
   const [modalVisible, setModalVisible] = useState(false);
+  const modalRef = useRef<HTMLDivElement | null>(null);
+  const backdropRef = useRef<HTMLDivElement | null>(null);
 
   useEffect(() => {
     setModalVisible(isDataPresented);
@@ -16,14 +24,44 @@ export default function Modal(props: ModalProps) {
     setModalVisible(false);
   };
 
+  const modalState = useAnimatedModalState(
+    modalVisible,
+    modalRef,
+    backdropRef,
+    onClosed
+  );
+
   return (
     <ModalProvider closeModal={onRequestClose}>
-      <ModalInner
-        isOpen={modalVisible}
-        onRequestClose={onRequestClose}
-        onClosed={onClosed}
-        {...rest}
-      />
+      {createPortal(
+        <div
+          ref={backdropRef}
+          className={clsx(
+            "modal-backdrop",
+            modalState !== "closed" ? "modal-visible" : "modal-not-visible"
+          )}
+          onClick={onRequestClose}
+        >
+          <div
+            ref={modalRef}
+            className="modal overflow-auto"
+            onClick={(e) => {
+              e.stopPropagation();
+            }}
+            {...rest}
+          >
+            <div className="modal-header-wrapper">
+              <div className="modal-header">
+                <h1>{title}</h1>
+                <X className="modal-header-exit-button" onClick={onRequestClose} />
+              </div>
+              <h2>{subtitle}</h2>
+            </div>
+            {children}
+          </div>
+        </div>,
+        document.body
+      )}
     </ModalProvider>
   );
 }
